pyproject.toml
[project]
name = "alrusdi-ru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "django-tinymce>=5.0.0",
    "gunicorn>=25.1.0",
    "python-dotenv>=1.2.1",
]
